# .well-known/openapi.yaml
openapi: 3.0.1
info:
  title: Auto Annotation
  description: A plugin that automatically annotates entity or classfy the text user input based types user specified.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /classification/{username}:
    post:
      operationId: getClassfications
      summary: Get the text classfication based on user input types
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/classificationRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getClassificationResponse'
  /entityextract/{username}:
    post:
      operationId: getEntities
      summary: Get the entity name and position based on user input types
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entityextractRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getEntityextractResponse'
components:
  schemas:
    classificationRequest:
      type: object
      properties:
        text:
          type: string
          description: The input text user want to classify
          required: true
        types:
          type: string
          description: The types user specified
          required: true
    getClassificationResponse:
      type: object
      properties:
        classes:
          type: string
          description: output class based on type provided by user
    entityextractRequest:
      type: object
      properties:
        text:
          type: string
          description: The input text user want to extract
          required: true
        types:
          type: string
          description: The types user specified
          required: true
    getEntityextractResponse:
      type: object
      properties:
        classes:
          type: array
          items:
            type: string
          description: array of eneity and their types